postgis 3.4.0 macos update

+++ postgis/3.4.0/macos/install.sh
@@ -2,27 +2,27 @@ FOLDER_NAME=postgis
 VERSION=3.4.0
 
 POSTGRES_FOLDER_NAME=postgres
-POSTGRES_VERSION=15.4
+POSTGRES_VERSION=16.0
 
-LIBXML_VERSION=2.10.4
+LIBXML_VERSION=2.11.5
 LIBXML_FOLDER_NAME=libxml2
 
 GEOS_VERSION=3.12.0
 GEOS_FOLDER_NAME=geos
 
-PROJ_VERSION=9.2.1
+PROJ_VERSION=9.3.0
 PROJ_FOLDER_NAME=proj
 
 PKG_CONFIG_VERSION=0.29.2
 PKG_CONFIG_FOLDER_NAME="pkg-config"
 
-SQLITE_VERSION=3.42.0
+SQLITE_VERSION=3.43.1
 SQLITE_FOLDER_NAME=sqlite3
 
-LIBTIFF_VERSION=4.5.1
+LIBTIFF_VERSION=4.6.0
 LIBTIFF_FOLDER_NAME=libtiff
 
-CURL_VERSION=8.2.1
+CURL_VERSION=8.3.0
 CURL_FOLDER_NAME=curl
 
 GETTEXT_FOLDER_NAME=gettext
@@ -41,12 +41,24 @@ if [ ! -d "$HOME/programs" ]; then
 	mkdir "$HOME/programs"
 fi
 
+if [ ! -d "$HOME/logs" ]; then
+	mkdir "$HOME/logs"
+fi
+
 if [ ! -d "$HOME/sources/$FOLDER_NAME" ]; then
 	mkdir "$HOME/sources/$FOLDER_NAME"
 fi
 
 if [ ! -d "$HOME/programs/$FOLDER_NAME" ]; then
 	mkdir "$HOME/programs/$FOLDER_NAME"
+fi
+
+if [ ! -d "$HOME/logs/$FOLDER_NAME" ]; then
+	mkdir "$HOME/logs/$FOLDER_NAME"
+fi
+
+if [ ! -d "$HOME/logs/$FOLDER_NAME/$VERSION" ]; then
+	mkdir "$HOME/logs/$FOLDER_NAME/$VERSION"
 fi
 
 if [ ! -d "$HOME/programs/$FOLDER_NAME/$VERSION" ]; then
@@ -83,24 +95,35 @@ if [ ! -d "$HOME/programs/$FOLDER_NAME/$VERSION" ]; then
 
 	cd $HOME/sources/$FOLDER_NAME
 
-	wget -q "https://postgis.net/stuff/postgis-$VERSION.tar.gz"
-	tar -xvf "postgis-$VERSION.tar.gz"
+	printf "${bold}${yellow}Installing $FOLDER_NAME${clear}\n"
+
+	printf "\t${bold}${green}Downloading source code${clear}\n"
+	wget -q --show-progress "https://postgis.net/stuff/postgis-$VERSION.tar.gz"
+	printf "\t${bold}${green}Extracting source code${clear}\n"
+	tar -xf "postgis-$VERSION.tar.gz"
 	mv "postgis-$VERSION" $VERSION
 	cd $VERSION
-	./configure --prefix=$HOME/programs/$POSTGRES_FOLDER_NAME/$POSTGRES_VERSION --with-projdir=$HOME/programs/$PROJ_FOLDER_NAME/$PROJ_VERSION --with-geosconfig=$HOME/programs/$GEOS_FOLDER_NAME/$GEOS_VERSION/bin/geos-config --without-protobuf --without-raster --with-gettext=$HOME/programs/$GETTEXT_FOLDER_NAME/$GETTEXT_VERSION --with-libintl-prefix=$HOME/programs/$GETTEXT_FOLDER_NAME/$GETTEXT_VERSION --disable-rpath
-	make
-	sudo make install
+	printf "\t${bold}${green}Configuring${clear}\n"
+	./configure --help > $HOME/logs/$FOLDER_NAME/$VERSION/configureHelp.txt 2>&1
+	./configure --prefix=$HOME/programs/$POSTGRES_FOLDER_NAME/$POSTGRES_VERSION --with-projdir=$HOME/programs/$PROJ_FOLDER_NAME/$PROJ_VERSION --with-geosconfig=$HOME/programs/$GEOS_FOLDER_NAME/$GEOS_VERSION/bin/geos-config --without-protobuf --without-raster --with-gettext=$HOME/programs/$GETTEXT_FOLDER_NAME/$GETTEXT_VERSION --with-libintl-prefix=$HOME/programs/$GETTEXT_FOLDER_NAME/$GETTEXT_VERSION --disable-rpath > $HOME/logs/$FOLDER_NAME/$VERSION/configureOutput.txt 2>&1
+	printf "\t${bold}${green}Making${clear}\n"
+	make > $HOME/logs/$FOLDER_NAME/$VERSION/makeOutput.txt 2>&1
+	printf "\t${bold}${green}Installing${clear}\n"
+	echo $USER_PASSWORD | sudo -S -p '' make install > $HOME/logs/$FOLDER_NAME/$VERSION/installOutput.txt 2>&1
 
-	cd $HOME/programs/$FOLDER_NAME/$VERSION
-	sudo chown -R $(whoami) .
+	if [ -e "$HOME/programs/$POSTGRES_FOLDER_NAME/$POSTGRES_VERSION/lib/postgis-3.dylib" ]; then
+		cd $HOME/programs/$POSTGRES_FOLDER_NAME/$POSTGRES_VERSION
+		echo $USER_PASSWORD | sudo -S -p '' chown -R $(whoami) .
 
-	install_name_tool -change @rpath/libgeos_c.1.dylib $HOME/programs/$GEOS_FOLDER_NAME/$GEOS_VERSION/lib/libgeos_c.1.dylib $HOME/programs/$POSTGRES_FOLDER_NAME/$POSTGRES_VERSION/lib/postgis-3.so
-	install_name_tool -change @rpath/libproj.25.dylib $HOME/programs/$PROJ_FOLDER_NAME/$PROJ_VERSION/lib/libproj.25.dylib $HOME/programs/$POSTGRES_FOLDER_NAME/$POSTGRES_VERSION/lib/postgis-3.so
-	install_name_tool -change @rpath/libgeos.3.12.0.dylib $HOME/programs/$GEOS_FOLDER_NAME/$GEOS_VERSION/lib/libgeos.3.12.0.dylib $HOME/programs/$GEOS_FOLDER_NAME/$GEOS_VERSION/lib/libgeos_c.1.18.0.dylib
+		install_name_tool -change @rpath/libgeos_c.1.dylib $HOME/programs/$GEOS_FOLDER_NAME/$GEOS_VERSION/lib/libgeos_c.1.dylib $HOME/programs/$POSTGRES_FOLDER_NAME/$POSTGRES_VERSION/lib/postgis-3.dylib
+		install_name_tool -change @rpath/libproj.25.dylib $HOME/programs/$PROJ_FOLDER_NAME/$PROJ_VERSION/lib/libproj.25.dylib $HOME/programs/$POSTGRES_FOLDER_NAME/$POSTGRES_VERSION/lib/postgis-3.dylib
+		install_name_tool -change @rpath/libgeos.3.12.0.dylib $HOME/programs/$GEOS_FOLDER_NAME/$GEOS_VERSION/lib/libgeos.3.12.0.dylib $HOME/programs/$GEOS_FOLDER_NAME/$GEOS_VERSION/lib/libgeos_c.1.18.0.dylib
 
-	cd $HOME/sources/$FOLDER_NAME
-	rm -rf $VERSION
-	rm "postgis-$VERSION.tar.gz"
+		printf "\t${bold}${green}Clearing${clear}\n"
+		cd $HOME/sources/$FOLDER_NAME
+		rm -rf $VERSION
+		rm "postgis-$VERSION.tar.gz"
+	fi
 fi
 
 cd $HOME/install-files
